frontend/api_agent/requirements.txt
Flask
flask-cors
python-dotenv
groq
google-generativeai
matplotlib
PyYAML
langchain>=0.0.200
